+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to Scrawler will be documented in this file.
 
 ## [Unreleased]
 ### Added
+- Added `getEntity` and `setEntity` methods to the `ResultWriterInterface` to
+  set the entity context for the result writer to use.
 - Added `initializeResultWrites` method to the `ResultWriterInterface` to allow
   hooking with any kind of preparation logic for every result writer. Refer to
   phpDoc/documentation to learn more.